Update cpp-files.sh
Added a condition that will check whether the directory $program_name 
already exists. If it exists then it gives an error because we 
won't want to add files in a directory that already exits 
(may be user is using it for some other purpose). I hope you like that 
change and include it in your script. It's a great script, we can
enhance it and make a script that works on design patterns just write what
pattern you want to use and files are created in same manner. 